+++ requirements/base.txt
@@ -8,9 +8,9 @@ appdirs==1.4.4
     # via fs
 asgiref==3.8.1
     # via django
-boto3==1.35.29
+boto3==1.35.34
     # via fs-s3fs
-botocore==1.35.29
+botocore==1.35.34
     # via
     #   boto3
     #   s3transfer
