+++ requirements/dev.txt
@@ -19,7 +19,7 @@ asgiref==3.8.1
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
     #   django
-astroid==3.3.4
+astroid==3.3.5
     # via
     #   -r requirements/quality.txt
     #   pylint
@@ -29,18 +29,18 @@ binaryornot==0.4.4
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
     #   cookiecutter
-boto3==1.35.29
+boto3==1.35.34
     # via
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
     #   fs-s3fs
-botocore==1.35.29
+botocore==1.35.34
     # via
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
     #   boto3
     #   s3transfer
-build==1.2.2
+build==1.2.2.post1
     # via
     #   -r requirements/pip-tools.txt
     #   pip-tools
@@ -357,7 +357,7 @@ requests==2.32.3
     #   -r requirements/test.txt
     #   cookiecutter
     #   xblock-sdk
-rich==13.8.1
+rich==13.9.2
     # via
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
@@ -404,11 +404,11 @@ tomlkit==0.13.2
     # via
     #   -r requirements/quality.txt
     #   pylint
-tox==4.20.0
+tox==4.21.2
     # via
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
-types-python-dateutil==2.9.0.20240906
+types-python-dateutil==2.9.0.20241003
     # via
     #   -r requirements/quality.txt
     #   -r requirements/test.txt
