+++ requirements/doc.txt
@@ -30,16 +30,16 @@ binaryornot==0.4.4
     # via
     #   -r requirements/test.txt
     #   cookiecutter
-boto3==1.35.29
+boto3==1.35.34
     # via
     #   -r requirements/test.txt
     #   fs-s3fs
-botocore==1.35.29
+botocore==1.35.34
     # via
     #   -r requirements/test.txt
     #   boto3
     #   s3transfer
-build==1.2.2
+build==1.2.2.post1
     # via -r requirements/doc.in
 cachetools==5.5.0
     # via
@@ -292,7 +292,7 @@ restructuredtext-lint==1.4.0
     # via doc8
 rfc3986==2.0.0
     # via twine
-rich==13.8.1
+rich==13.9.2
     # via
     #   -r requirements/test.txt
     #   cookiecutter
@@ -348,11 +348,11 @@ text-unidecode==1.3
     # via
     #   -r requirements/test.txt
     #   python-slugify
-tox==4.20.0
+tox==4.21.2
     # via -r requirements/test.txt
 twine==5.1.1
     # via -r requirements/doc.in
-types-python-dateutil==2.9.0.20240906
+types-python-dateutil==2.9.0.20241003
     # via
     #   -r requirements/test.txt
     #   arrow
