+++ requirements/quality.txt
@@ -16,7 +16,7 @@ asgiref==3.8.1
     # via
     #   -r requirements/test.txt
     #   django
-astroid==3.3.4
+astroid==3.3.5
     # via
     #   pylint
     #   pylint-celery
@@ -24,11 +24,11 @@ binaryornot==0.4.4
     # via
     #   -r requirements/test.txt
     #   cookiecutter
-boto3==1.35.29
+boto3==1.35.34
     # via
     #   -r requirements/test.txt
     #   fs-s3fs
-botocore==1.35.29
+botocore==1.35.34
     # via
     #   -r requirements/test.txt
     #   boto3
@@ -266,7 +266,7 @@ requests==2.32.3
     #   -r requirements/test.txt
     #   cookiecutter
     #   xblock-sdk
-rich==13.8.1
+rich==13.9.2
     # via
     #   -r requirements/test.txt
     #   cookiecutter
@@ -302,9 +302,9 @@ text-unidecode==1.3
     #   python-slugify
 tomlkit==0.13.2
     # via pylint
-tox==4.20.0
+tox==4.21.2
     # via -r requirements/test.txt
-types-python-dateutil==2.9.0.20240906
+types-python-dateutil==2.9.0.20241003
     # via
     #   -r requirements/test.txt
     #   arrow
