+++ requirements/test.txt
@@ -16,11 +16,11 @@ asgiref==3.8.1
     #   django
 binaryornot==0.4.4
     # via cookiecutter
-boto3==1.35.29
+boto3==1.35.34
     # via
     #   -r requirements/base.txt
     #   fs-s3fs
-botocore==1.35.29
+botocore==1.35.34
     # via
     #   -r requirements/base.txt
     #   boto3
@@ -182,7 +182,7 @@ requests==2.32.3
     # via
     #   cookiecutter
     #   xblock-sdk
-rich==13.8.1
+rich==13.9.2
     # via cookiecutter
 s3transfer==0.10.2
     # via
@@ -207,9 +207,9 @@ stevedore==5.3.0
     # via code-annotations
 text-unidecode==1.3
     # via python-slugify
-tox==4.20.0
+tox==4.21.2
     # via -r requirements/test.in
-types-python-dateutil==2.9.0.20240906
+types-python-dateutil==2.9.0.20241003
     # via arrow
 urllib3==2.2.3
     # via
