+++ xblocks_contrib/html/html.py
@@ -12,7 +12,7 @@ resource_loader = ResourceLoader(__name__)
 
 
 # This Xblock is just to test the strucutre of xblocks-contrib
-@XBlock.needs('i18n')
+@XBlock.needs("i18n")
 class HtmlBlock(XBlock):
     """
     TO-DO: document what your XBlock does.
@@ -28,6 +28,10 @@ class HtmlBlock(XBlock):
         help="A simple counter, to show something happening",
     )
 
+    is_extracted = (
+        True  # Indicates that this XBlock has been extracted from edx-platform.
+    )
+
     def resource_string(self, path):
         """Handy helper for getting resources from our kit."""
         return files(__package__).joinpath(path).read_text(encoding="utf-8")
@@ -41,13 +45,15 @@ class HtmlBlock(XBlock):
             pass  # TO-DO: do something based on the context.
 
         frag = Fragment()
-        frag.add_content(resource_loader.render_django_template(
-            'templates/html.html',
-            {
-                'count': self.count,
-            },
-            i18n_service=self.runtime.service(self, 'i18n')
-        ))
+        frag.add_content(
+            resource_loader.render_django_template(
+                "templates/html.html",
+                {
+                    "count": self.count,
+                },
+                i18n_service=self.runtime.service(self, "i18n"),
+            )
+        )
 
         frag.add_css(self.resource_string("static/css/html.css"))
         frag.add_javascript(self.resource_string("static/js/src/html.js"))
