+++ xblocks_contrib/lti/tests/test_lti.py
@@ -1,27 +1,26 @@
 """
-Tests for ProblemBlock
+Tests for LTIBlock
 """
-
 
 from django.test import TestCase
 from xblock.fields import ScopeIds
 from xblock.test.toy_runtime import ToyRuntime
 
-from xblocks_contrib import ProblemBlock
+from xblocks_contrib import LTIBlock
 
 
-class TestProblemBlock(TestCase):
-    """Tests for ProblemBlock"""
+class TestLTIBlock(TestCase):
+    """Tests for LTIBlock"""
 
     def test_my_student_view(self):
         """Test the basic view loads."""
         scope_ids = ScopeIds("1", "2", "3", "4")
-        block = ProblemBlock(ToyRuntime(), scope_ids=scope_ids)
+        block = LTIBlock(ToyRuntime(), scope_ids=scope_ids)
         frag = block.student_view()
         as_dict = frag.to_dict()
         content = as_dict["content"]
         self.assertIn(
-            "ProblemBlock: count is now",
+            "LTIBlock: count is now",
             content,
             "XBlock did not render correct student view",
         )
