+++ xblocks_contrib/problem/tests/test_problem.py
@@ -1,27 +1,26 @@
 """
-Tests for WordCloudBlock
+Tests for ProblemBlock
 """
-
 
 from django.test import TestCase
 from xblock.fields import ScopeIds
 from xblock.test.toy_runtime import ToyRuntime
 
-from xblocks_contrib import WordCloudBlock
+from xblocks_contrib import ProblemBlock
 
 
-class TestWordCloudBlock(TestCase):
-    """Tests for WordCloudBlock"""
+class TestProblemBlock(TestCase):
+    """Tests for ProblemBlock"""
 
     def test_my_student_view(self):
         """Test the basic view loads."""
         scope_ids = ScopeIds("1", "2", "3", "4")
-        block = WordCloudBlock(ToyRuntime(), scope_ids=scope_ids)
+        block = ProblemBlock(ToyRuntime(), scope_ids=scope_ids)
         frag = block.student_view()
         as_dict = frag.to_dict()
         content = as_dict["content"]
         self.assertIn(
-            "WordCloudBlock: count is now",
+            "ProblemBlock: count is now",
             content,
             "XBlock did not render correct student view",
         )
