+++ xblocks_contrib/video/tests/test_video.py
@@ -2,7 +2,6 @@
 Tests for VideoBlock
 """
 
-
 from django.test import TestCase
 from xblock.fields import ScopeIds
 from xblock.test.toy_runtime import ToyRuntime
