+++ xblocks_contrib/word_cloud/tests/test_word_cloud.py
@@ -1,27 +1,26 @@
 """
-Tests for HtmlBlock
+Tests for WordCloudBlock
 """
-
 
 from django.test import TestCase
 from xblock.fields import ScopeIds
 from xblock.test.toy_runtime import ToyRuntime
 
-from xblocks_contrib import HtmlBlock
+from xblocks_contrib import WordCloudBlock
 
 
-class TestHtmlBlock(TestCase):
-    """Tests for HtmlBlock"""
+class TestWordCloudBlock(TestCase):
+    """Tests for WordCloudBlock"""
 
     def test_my_student_view(self):
         """Test the basic view loads."""
         scope_ids = ScopeIds("1", "2", "3", "4")
-        block = HtmlBlock(ToyRuntime(), scope_ids=scope_ids)
+        block = WordCloudBlock(ToyRuntime(), scope_ids=scope_ids)
         frag = block.student_view()
         as_dict = frag.to_dict()
         content = as_dict["content"]
         self.assertIn(
-            "HtmlBlock: count is now",
+            "WordCloudBlock: count is now",
             content,
             "XBlock did not render correct student view",
         )
